pyproject.toml
[tool.poetry]
name = "airsync"
version = "0.3.0"
description = ""
authors = ["Daniel Roziecki <emssik@emssik.eu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pyairtable = "^2.3.6"
psycopg2 = "^2.9.10"
pyyaml = "^6.0.2"
pytest = "^7.0.0"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
